+++ packages/android_intent/android/src/main/java/io/flutter/plugins/androidintent/AndroidIntentPlugin.java
@@ -46,6 +46,8 @@ public class AndroidIntentPlugin implements MethodCallHandler {
         return Settings.ACTION_SETTINGS;
       case "action_location_source_settings":
         return Settings.ACTION_LOCATION_SOURCE_SETTINGS;
+      case "action_application_details_settings":
+        return Settings.ACTION_APPLICATION_DETAILS_SETTINGS;
       default:
         return action;
     }
@@ -125,6 +127,9 @@ public class AndroidIntentPlugin implements MethodCallHandler {
     Intent intent = new Intent(action);
     if (mRegistrar.activity() == null) {
       intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+    }
+    if (call.argument("flag") != null) {
+      intent.addFlags((Integer) call.argument("flags"));
     }
     if (call.argument("category") != null) {
       intent.addCategory((String) call.argument("category"));